byte

+++ pacman/src/main/java/de/amr/games/pacman/game/worlds/PacManGameWorld.java
@@ -14,7 +14,7 @@ public interface PacManGameWorld {
 
 	public static final byte SPACE = 0, WALL = 1, PILL = 2, ENERGIZER = 3, DOOR = 4, TUNNEL = 5;
 
-	public static final int TS = 8, HTS = 4;
+	public static final byte TS = 8, HTS = 4;
 
 	public static int t(int nTiles) {
 		return nTiles * TS;
